requirements.txt
smolagents
pandas
openmeteo_requests
python-dotenv
stramlit